split install state CA
